Revert "Combined appointment and patient add and delete, added JUnit as well"
This reverts commit b760ef1de9a060573275592deced31866b40a41e.
