+++ src/main/java/Patient.java
@@ -1,5 +1,3 @@
-package MainFunction;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,13 +20,28 @@ public class Patient {
         this.medicalHistory = new ArrayList<>();
     }
 
-    public String getName() {return name;}
+    public String getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDob() {
+        return dob;
+    }
+
+    public String getContactInfo() {
+        return contactInfo;
+    }
+
+    public List<String> getMedicalHistory() {
+        return medicalHistory;
+    }
 
     @Override
     public String toString() {
-        return String.format(
-                "MainFunction.Patient ID: %s\n   Name: %s\n   Date of Birth: %s\n   Gender: %s\n   Address: %s\n   Contact: %s",
-                id, name, dob, gender, address, contactInfo
-        );
+        return "Patient ID: " + id + ", Name: " + name + ", Date of Birth: " + dob + ", Contact: " + contactInfo;
     }
 }
